Resolve QUARC license not recognized in MATLAB on Apple Silicon
Issue is that MATLAB (at the moment R2023a) running on Apple Silicon is still using Rosetta2, thus it still requires an emulated Intel arch. Thus the license registration tool needs to run twice, one to register with mac64 (for real-time code) and one for maci64 (for QUARC in MATLAB).